refactor: server/database
